FCM push notifications (#14)
This replaces background notifications with FCM notifications which are
much more reliable (I can't get the background fetches to fire
regularly).

Now a GCP function reads the existing users from a Firestore collection
every few minutes, checks whether any alarm is due and if so, fetches
the weather for the user's location, determines the right clothes to
wear and sends a push notification with the recommendation. This
required pulling out some shared code into a new `shared` directory
which is depended on by both the GCP function and the app.